refactor: move console platform code
