+++ packages/log/ts/platform/index.browser.ts
