+++ packages/log/ts/platform/index.ts
@@ -1,4 +1,3 @@
-// istanbul ignore next
 export function semverGt(versionString: string, version: [number, number, number]) {
   const actual = versionString.split('.').reverse().reduce((p, v, i) => {
     p += Number.parseInt(v, 10) * Math.pow(100, i)
@@ -6,4 +5,11 @@ export function semverGt(versionString: string, version: [number, number, number
   }, 0)
 
   return actual > (version[0] * 10000 + version[1] * 100 + version[2])
+}
+
+export function isConsoleDebugAvailable() {
+  // without this, systemJs will complain `process is not defined`
+  if (!global.process) return true
+  const versionString = process.version.startsWith('v') ? process.version.slice(1) : process.version
+  return semverGt(versionString, [9, 3, 0])
 }
